:gift:Feat:: base64 str을  url로 전환하는 함수 구현
